Solved the sudoku problem 96
This was more brute force then I wanted it to be. One problem is that I
believed that all the sudoku puzzles could be solved with logic but some
of them seemed to not be able to. I had to change it to a test check
method after eliminating with some logical checks.